# openapi.yaml
openapi: 3.1.0
info:
  title: Convocate API
  version: 0.1.0
  description: |
    Convocate is an agentic software-engineering platform: it provisions compute
    **Nodes**, runs **Agent-containers** (Claude CLI in permission-bypass mode) on
    them, decomposes a natural-language **Specification** into a **Project Board**
    of Cards/Containers wired into execution DAGs, executes those cards against
    agents to produce code, and manages the resulting **Repositories**, **Pull
    Requests** and CI/CD through to merge.

    ## Conventions
    - All endpoints live under `/api/v1/<applet_shortname>/...`, where
      `applet_shortname` is the 2–4 letter abbreviation for the owning applet:
      `nmgr` (Node Manager), `amgr` (Agent Manager), `pb` (Project Board),
      `ide` (Code IDE), `repo` (Repo Manager), `ac` (Access Control),
      `sup` (Support Tool), plus `auth` for the session/identity service.
    - JSON request/response bodies; `application/json` unless noted.
    - Authentication via Bearer JWT (see `securitySchemes.bearerAuth`).
    - Mutating endpoints are authorized by RBAC role (see `x-required-role`);
      a principal holding the `admin` role implicitly satisfies every check.
    - List endpoints are cursor/offset paginated and return `Page<T>`.
    - Timestamps are RFC 3339 / ISO 8601 UTC strings.
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: convocate.local
security:
  - bearerAuth: []
...
paths:
...
  /sup/ticket:
    get:
      tags: [sup]
      summary: List support tickets
      x-required-role: support-view
      parameters:
        - { $ref: '#/components/parameters/offset' }
        - { $ref: '#/components/parameters/limit' }
      responses:
        '200':
          description: Page of tickets
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/Ticket' } }
    post:
      tags: [sup]
      summary: Open a support ticket
      x-required-role: support-view
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TicketInput' } } }
      responses:
        '201': { description: Ticket, content: { application/json: { schema: { $ref: '#/components/schemas/Ticket' } } } }
  /sup/ticket/{ticketId}:
    parameters: [ { name: ticketId, in: path, required: true, schema: { type: string } } ]
    get:
      tags: [sup]
      summary: Get ticket detail
      x-required-role: support-view
      responses:
        '200': { description: Ticket, content: { application/json: { schema: { $ref: '#/components/schemas/Ticket' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [sup]
      summary: Update a support ticket (status, priority, body)
      x-required-role: support-view
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TicketInput' } } }
      responses:
        '200': { description: Updated ticket, content: { application/json: { schema: { $ref: '#/components/schemas/Ticket' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /sup/doc:
    get:
      tags: [sup]
      summary: List documentation articles
      x-required-role: support-view
      parameters:
        - { $ref: '#/components/parameters/offset' }
        - { $ref: '#/components/parameters/limit' }
      responses:
        '200':
          description: Page of articles
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/DocArticle' } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    offset: { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
    limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 25 } }
...
  responses:
...
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: validation_failed }
        message: { type: string }
        details:
          type: array
          items: { type: object, properties: { field: { type: string }, issue: { type: string } } }
    Page:
      type: object
      properties:
        offset: { type: integer }
        limit: { type: integer }
        total: { type: integer }

    # ---- status ----
...
    TicketStatus: { type: string, enum: [open, in-progress, resolved, closed] }
    TicketPriority: { type: string, enum: [low, medium, high] }
    Ticket:
      type: object
      properties:
        id: { type: string }
        subject: { type: string }
        status: { $ref: '#/components/schemas/TicketStatus' }
        priority: { $ref: '#/components/schemas/TicketPriority' }
        body: { type: string }
        updatedAt: { type: string, format: date-time }
    TicketInput:
      type: object
      required: [subject]
      properties:
        subject: { type: string }
        priority: { $ref: '#/components/schemas/TicketPriority' }
        body: { type: string }
    DocArticle:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        slug: { type: string }